Create RangePartition.java
Alan and Barbara suddenly felt like playing with numbers. Alan chooses a non-empty subset from the set of first N positive integers (1,2,…,N). Barbara takes the rest of the numbers (if any) from the set. And then they both calculate the sum of the elements in their respective sets.

Alan believes in a magic ratio, which is X:Y. Hence, Alan wants to choose the subset in such a way that the ratio between the sum of Alan's subset and the sum of Barbara's subset is exactly X:Y.

Can you help Alan to choose a subset that can achieve the desired ratio?